Create README.md
How identity layering improves data flow

The layered approach for security is dynamic. Will be continue to develop the layered digital identity and identity in depth.

Back in the day, a new idea was floated called “layered security.” It was a model that helped an organization plan out how to secure across all the touchpoints of the business; each layer protecting against a threat. Layered security evolved into the more holistic “Defense in Depth.” This new model was based on the philosophy of the whole being greater than the sum of the parts. It is a good model to follow and one that I believe the digital identity space can replicate. Here's how.

The word ‘identity’ conjures up an emotional response. And why not? It refers to our very being. But the actual use of a digital identity is not really about me, myself, and I. It’s about me doing stuff online (or offline as the use case may be); digital identity in the consumer space is about people and data. Identity, in and of itself, is not the issue — it is how we present data to do a job that matters. Now, this data, which may represent some or many aspects of personally identifying information (PII), are best utilized on a need to know basis. If you work outwards from that premise, you also start the process to design using privacy as a remit.

People and data doing jobs need conduits or facilitators to work with and across. This is where the idea of ‘identity layering’ comes in. And, it is an architectural concept that is the key to making consumer IAM more than just a marketing tool.

Identity data layering allows the PII to flow but under the control of the user within the constraints of rules.

#data #bigdata #digitalidentity #sovereignty #cybersafety #privacy #tech #blockchain #viral #digitalsecurity #share